requirements.txt
# Blockchain & Algorand
py-algorand-sdk==2.6.0
pyteal==0.24.0

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# AI & ML
opencv-python==4.8.1.78
pytesseract==0.3.10
scikit-learn==1.3.2
numpy==1.26.2
pandas==1.5.3
Pillow==10.1.0

# Cryptography & Security
cryptography==41.0.7
pycryptodome==3.19.0

# Database
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
